requirements.txt
gradio>=4.0.0
folium>=0.15.0
anthropic>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.0.0